(GH-371) Enable self-service mode
When using self-service installs, a feature in Chocolatey for Business,
it provides a background agent that actually performs the
installations. Chocolatey itself handles passing through the proper
information to the background service so that installations can occur.
The GUI should notice the feature is switched on and adjust to allow
non-administrators to run certain commands in the GUI without
attempting to elevate.

Original commit from @RichiCoder1.