(GH-372) Dynamically select open port for subprocess
Currently Chocolatey GUI will attempt to acquire a static port, and simply fail if this port is already in use. This change has GUI check for an available port, and only fail if it can't find one open in an available range. 

(This was already already implemented in Subprocess, it was just never brought over to GUI side).