(GH-371) Ensure config is properly set
When calling into chocolatey.lib, if the Container is accessed
prior to having the configuration all set, any additonal components
that Chocolatey attempts to register will be ignored. This causes
issues with using Chocolatey and registering licensed configuration
and components. Ensure that Chocolatey does get set appropriately
by making a call to do that early in the process load.