(GH-371) Enable self-service mode (#451)
* (GH-371) Enable self-service mode

When using self-service installs, a feature in Chocolatey for Business,
it provides a background agent that actually performs the
installations. Chocolatey itself handles passing through the proper
information to the background service so that installations can occur.
The GUI should notice the feature is switched on and adjust to allow
non-administrators to run certain commands in the GUI without
attempting to elevate.

Original commit from @RichiCoder1.

* (maint) capture connection closed error

There are errors where a connection was closed during an operation,
like when the application is being closed when running a query. Capture
those and allow the application to exit peacefully.

Original commit from @RichiCoder1.

* (maint) Add Pin for German resources

* (GH-425) New source settings

- Added necessary UI elements
- Updated model classes with new properties

* (GH-371) Ensure config is properly set

When calling into chocolatey.lib, if the Container is accessed
prior to having the configuration all set, any additonal components
that Chocolatey attempts to register will be ignored. This causes
issues with using Chocolatey and registering licensed configuration
and components. Ensure that Chocolatey does get set appropriately
by making a call to do that early in the process load.

* (maint) upgrade to Chocolatey 0.10.7

* Ensure we don't elevate when operating through a background agent.