(GH-495) modernize install script
The chocolateyInstall.ps1 script did not contain the `/norestart`
option, which would cause an install to possibly reboot a machine,
depending on whether the machine was in or set to a pending reboot.
Installing/upgrading the package should not cause an upgrade.

Add valid exit codes for an MSI and add MSI logging in case any errors
occur.