(GH-496) Ensure WixProj does not rebuild GUI
The GUI project was previously a referenced project, so if changes were
detected it would rebuild ChocolateyGui project. Typically once the
main exe ChocolateyGui.exe was authenticode signed, this would result
in a triggered rebuild, removing any authenticode signatures. Instead
of depending on a project, look for the files where the project output
will be instead of depending on a project reference.