(build) Fix sign criteria for null paths
There was a thought that FileExists was robust enough to handle nulls
or empty strings and return false (because of course a null file
doesn't exist), but it appears that it simply does no validation and
allows the null value check against File.Exists, which errors on
missing paths. Check to be sure the value is empty prior to sending it
to the helper method.