(GH-629) Separate app.manifest
By default highestAvailable is desired for UAC prompting behavior, but
sometimes there are groups that provide some level of elevation
(Network Configuration Operators Group) that also mean that UAC will
prompt. This is undesirable behavior in shops that might want to use
self-service and have Chocolatey GUI run as the user that invoked
without any elevation being requested. To allow this behavior,
app.manifest can not be embedded, but must sit next to
ChocolateyGUI.exe so that it can be changed for those situations that
require it by those users.