(GH-603) Fix bug when updating source information
When we changed to getting sources directly from Chocolatey.Lib, rather
reading directly from the config file, we introduced a bug.  Other parts
of Chocolatey GUI expect to have all information about a source
available to it, i.e. UserName and Password of an authenticated feed.
However, this information isn't available when using Chocolatey.Lib to
get the sources.  Instead, still read the sources from the config file
as before, but filter this list, based on the sources that are retrieved
from Chocolatey.Lib.  This is a duplication of effort, however, it does
mean that the work of filtering the sources list remains where it needs
to be, i.e. in Chocolatey.Lib.