(GH-574) Add shimgen for ChocolateyGui.exe
To allow the usage of ChocolateyGui.exe directly from the Command Line.
Normally, this isn't required, however, new functionality allows the
application to work as both a command line application, and a GUI
application.  This is to allow setting on/off the features of Chocolatey
GUI.  The shimgen isn't "quite" ideal, in that the response from the
application, when running as a command line appears after the prompt
has returned, but this is better than having to wait for the GUI to
close before focus is returned to the command line.  This can always
be changed later if there is a greater call for it.