(GH-430) Fix missing Gallery link in details page
Initially, I thought the issue here was one where the information from
Chocolatey API didn't have everything that was needed compared to what
used to come from the oData feed, however, this wasn't the case.  In
the end, it was a mapper issue.  IPackage doesn't have a property
called GalleryDetailsUrl, but it does have one called ReportAbuseUrl.
That is why that one worked, but the other one didn't.  Adding an
additional map between DataServicePackage and Package is what was
needed to make all this work.  This concrete type has all the required
properties needed to display information in the UI.