(GH-430) Fix Package Size displaying incorrectly
There are times when, for whatever reason, information is missing for
a package.  For example, download count, etc.  This is also true of the
package size property, which can have a default value of -1.  In these
situations, the package size label, and associated text block shouldn't
be displayed.  To allow this, a new IsPackageSizeAvailable property has
been created, which is used as the Binding to the Visibility of the UI
elements that show this information.  This really only affects the
local package source, as the remote sources "should" have this
information available for display.