(GH-430) Add additional mapping between types
The fix that was added to correctly show the GalleryDetailsUrl actually
introduced a bug due to the different type for the Authors property on
the DataServicePackage type.  Here, Authors is a string, however, it is
expected that it is an IEnumerable<string>, and as a result, the string
is broken up into individual characters to make it an array.  Have added
explicit handling for this mapping to account for this difference.