(GH-525) Implement caching of Outdated Packages
Outdated Packages are now going to be serialized to a file in the users
appdata folder.  This is facilitated by a new command added to the GUI
which allows setting of a config variable to control how long the cache
should exist.

Whilst doing this work, the existing Feature Command was refactored to
use the new generic implmenetation of an ICommand interface, similar to
how Chocolatey itself works.

All "output" from these commands has been made localizable.