(GH-665) Don't allow direct editing of passwords
When editing a source via Chocolatey GUI, the source is first removed
and then re-added with the current information.  This is problematic
due to the fact that the encrypted password isn't first decrypted. As
a result, the encrypted password is then re-encrypted by Chocolatey, and
therefore invalid.

Now, the password is first decrypted before being saved again, and the
password values are no longer displayed to the user in the Sources
screen as there is no point, since they are encrypted.

Special handling has been added for the chocolatey.licensed source to
only allow enabling/disabling of the source as that is all Chocolatey
Licensed actually allows.