(GH-270) Ability to show additional package info
A request was made to allow additional information to be shown when
viewing a package, including the id of the package, and whether it is a
prerelease package.  This has been enabled as a feature which can be
turned on and off.  It isn't enabled all the time, as not everyone wants
this additional information to be displayed.