(GH-682) Refactor project structure
There is a single class that needs to be split into it's own project.
This class requires a reference to WindowsBase and PresentationCore,
which are NOT required in the current Common project.  Rather than
pull those dependencies into the Common library, create a Common.Windows
library, which will serve that purpose.