(GH-679) Make extension points for GUI Extension
So that different features can be lit up based on the presence of the
chocoalteygui.extension package, a number of changes had to be made
to the core of Chocolatey GUI.  Firstly, if the GUI Extension Assembly
exists, then any Modules that exist within it have to be loaded into the
current container.  With this in place, application logic and UI was
altered to use functionality from interfaces, so that this functionality
can be changed based on loading a new implementation of that interface.