(GH-704) Add tasks to strong name output
This involves finding the snk file to do the signing, either locally, or
defined in an environment variable. In addition, it is necessary to
also strong name some referenced assemblies, and this is done via the
Brutal.Dev.StrongNameSigner tool.  Only a subset of the referenced
assemblies are signed, as some already are, and trying to do all of them
led to very long build times.