(GH-708) New ReleaseOfficial build configuration
This will be used when preparing for an official release of Chocolatey,
to ensure that the correct strong name key is being used, and also that
the necessary checks are being done when loading the Chocolatey GUI
Licensed assembly into the Chocolatey process.  The
FORCE_CHOCOLATEY_OFFICIAL_KEY compilation symbol if used in
if/def statements to make this logic happen.