(build) Fix issue with build
The NuGet package was being created before the assemblies were being
signed, so had to alter the DAG to account for this.

In addition, found that the generated MSI wasn't including the files
that had been signed, so altered the wixproj to take these files from the
correct location.

Finally, updated the Configuration Manager to ensure that the WIX
configuration only builds the install project, as that is all that is
required.