(GH-779) Don't set Prefer32Bit to true
This setting was set to false for both Debug and Release Build
configurations, but for some reason on the ReleaseOfficial
configuration it was set to true.  As a result, the shipped version
of the exe was set to true.  During testing, it was clear that with
the setting enabled, it was indeed running as a 32 bit process, where
as Chocolatey itself was running as a 64 bit process.  Disabling this
setting meant it was allowed to run as a 64 bit process.