(GH-602) Add global configuration
This introduces the concept of a data.db file in the
C:/ProgramData/Chocolatey GUI folder, which will house the global
configuration for the application.

Calling chocolateyguicli feature or config with the --global flag will
set the data.db in this new location.  Without the --global flag, the
setting will be made in the users data.db file, as normal.

Then, the effective configuration, i.e. the current Global configuration
and any overridden user settings, is evaluated, and used as the current
configuration for the running application.