(GH-785) Remove ApplicationParameters class
Due to problems with using an assembly before it has been correctly
resolved, using the concept of a shared ApplicationParameters class
won't work going forward.  Instead, the required information needs to
be duplicated in both the Chocolatey GUI and the Chocolatey CLI
projects.  This may cause problems later with StyleCop when we come to
create a release build, but we will cross that bridge when we come to
it.