(GH-785) Add additonal project/nuget references
When loading the Chocolatey GUI licensed extension, there are a couple
of additional project/nuget refereces that need to be in place.  These
are not strictly "required" by the CLI project, however, switching to
the assembly resolver class means that these need to be in place.  This
will help with debugging of the CLI project when the Chocolatey GUI
licensed extension is in place.