(GH-785) Add assembly resolver to CLI and GUI
This will be responsible for ensuring that a version of the
ChocolateyGui.Common and ChocolateyGui.Common.Windows assemblies
are loaded into the current process.  As long as the Public Key of the
requested assembly matching what the host process has been compiled
with, then it will be returned.  The compiled version number does not
need to match.  This will mean that the Chocolatey GUI Extension will
work with multiple versions of Chocolatey GUI, in the same was as the
Chocolatey Extension works with multiple versions of Chocolatey.