(GH-807) Ensure that Local Only flag is false
When searching for packages from a remote source, ensure that the Local
Only flag is always set to false, otherwise, Chocolatey will return a
list of local packages, rather than those from the remote source. This
was discovered when clicking on a remote source while the list of
locally installed packages was still being populated. The search of the
remote source was still using some of the configuration from the query
of the locally installed packages.