(GH-808) Fix bug with view activation
In an earlier commit for this issue, the _firstLoad logic was removed,
as it was believed that this was no longer required, however, that is
not the case.  This logic is needed to ensure that when returning from
the package details view, that sources aren't incorrectly loaded again.

In addition, it was found that after the earlier changes were made, the
showing of the aggregated source also no longer worked.  Not entirely
sure how this ever worked, as there wasn't a message sent to update the
sources when this feature was activated.  This has now been added.