(#757) Allow buttons to be enabled/disabled
The decision was taken to allow a button to be enabled/disabled based on
current Chocolatey configuration (mainly when running in background
service mode). To allow for this, whilst still allowing a button to be
enabled/disabled based on application logic, a new behavior was
introduced to encapsulate this logic.

This will allow the Chocolatey GUI Licensed Extension to override the
AllowedCommandsService, to control which operations should be enabled
and disabled, based on the background service configuration.

Co-authored-by: punker76 <punkerat76@gmail.com>