(#820) Change default for tile view
These will now be enabled for both local and remote out of the box. This
is a breaking change, so taking the opportunity to rename the collection
to match the version of the application. This will allow for different
versions of the configuration per application version, but it will mean
that we may need to migrate settings going forward.