(#829) Ensure application starts without global db
There is a situation when the global configuration database doesn't
exist, for example, when Chocolatey GUI is installed by administrator
but never executed by that user. When running as a non-administrrator
user, the application would then crash.  This changes makes it possible
for the application to simply ignore the fact that the global database
doesn't exist, and instead, simply use the default settings, with no
attempt to then write any values back to the database.

In order to work correctly, there is an update required in the Chocolatey
GUI Licensed extension to make things work there as well.