(#832) Ensure to correctly dispose of database
If this isn't done, then a data-log.db file is created when a change is
made to the configuration databases, and it can be that these changes
are not persisted to the main data.db file.

In future versions, we may need to look at handling the usage of the
database(s) differently, and to _not_ put them into the IoC container,
but for now, explicitly disposing of each database when the application
exits, and when an exception occurs, seems like a good compromise.