(#838) Add ability to filter features/settings
For both Chocolatey GUI and Chocolatey. This will make it much easier
to find the setting that you are interested in changing. This has been
implemented by using an ICollectionView for each collection, and then
applying a filter method to each one, based on what is typed into the
search textbox.