(#770) Add ability to view install arguments
This allows the viewing of the contents of the .arguments file, which
is stored on a per package version, and contains the arguments that
were passed into the choco install command for that package.

This file is encrypted, and can't be viewed directly, but this allows
decryption of the contents for viewing.

Co-authored-by: steviecoaster <stephen@chocolatey.io>