(#882) Make use of OSS Export Command
Rather than use our own logic to generate a list of installed packages,
make use of the new Chocolatey CLI OSS command which does this for us.
At the minute, there is no way in Chocolatey GUI to control whether
version numbers are included or not, but this could come in a future
modification.  For now, provide the same functionality that was
included before with this button.

Also added a message to indicate that the operation had completely
successfully.

To provide this functionality, the IChocolateyService was extended to
include a new ExportPackages method, which simply runs the Chocolatey
Export command.