(#904) Wrap MessageBoxes with a hidden window
There appears to be a "feature" in WPF that prevents MessageBoxes from
displaying when there are no active windows. Many of our MessageBoxes
are to report errors, and many of them occur during startup. This
results in them not displaying as we haven't created any windows. This
works around that by creating a hidden window and closing it after the
MessageBox has been displayed.