(#885) Remove package parameters from nuspec file
This commit updates the Chocolatey Package to instead of embedding
the package parameter for the package to link to the documentation
page for Chocolatey GUI on https://docs.chocolatey.org instead.

This was needed as we are starting to hit the character limitation
specified for the description in packages (4 000 characters).