(#533) Update TranslationSource to accept parameters
Quite a few of the strings that are being used for
translations accepts parameters as such it would
be helpful if the translation source accepts an
undefined number of arguments instead of having
to call a manually a different formatting method.

This commit updates that translation source to accept
source parameters.