(#533) Update direct calls to Resources to use translation source
This commit updates the existing calls to `Resource.` static class to instead
make calls to our own custom `TranslationSource` class to ensure that we
are using the same logic for the translation strings everywhere in the code.

Additionally, there was a few places in the code that would not use the correct
language when calling the Resource class directly, switch to the translation source
fixes those.