(maint) Remove format support for install configuration description
This comit updates the configuration description when doing
advanced installation by removing the parts of the string
that need to be formatted with the value that have the
current configuration have as defaults.

Additional, the mention of which Chocolatey versions that
support the respective feature have been removed from
the view.