(#545) Make version editable on advanced install
This commit updates the code necessary to make the version
combobox editable while allowing the user to select the latest version,
or the version they opened the Advanced Installation window on.

Additionally, due to known issues with populating available versions for
the user (both on CCR, Nexus and Artifactory) fetching of remote versions
are disabled until further notice.