(#533) Fix remote sources being reloaded on language change
This commit fixes an issue where if the remote sources view was
selected before going into the settings view, when the language
is then changed it would reload the packages that was already
acquired which at some rare times could result in an endless
loop.

This fix changes that, to instead only reload the sources if the
remote sources view is active, and as thus checks for the sources
when the view gets activated the first time instead of during
the initialization phase.