(#533) Set current culture for WPF elements
This comit adds missing code that should have been set WPF
elements with the correct language code (and direction) when
a language is initially set, or when the language gets changed.

The right to left changes is not perfect, but until there is feedback
from users that uses such a locale it is the best we can do at the
moment.