(#545) Remove NumberOfPackageVersionsForSelection
This setting was added when there was a plan to include a filtered list
of available package versions when using the advanced installation
dialog. Due to the fact that fetching the available package versions is
known to be problematic, this feature has been pushed out to later
version. As a result, having this configuration setting doesn't make
any sense, and will likely lead to confusion if we leave it in.  We
will likely add this setting back in, if/when that functionality is
added in.