(#533) Handle null culture
During testing, it was found that when used in conjunction with
Chocolatey GUI Licensed Extension, it was possible for the initial
culture to be null.

The thought is that this is due to how the ShellViewModel is being
overriden in the Licensed Extension. It was found that there isn't an
issue with simply returning from the SetLanguage method, as everything
just reverts to default, until it is explicitly set.