(#923) Switch to using SolutionVersion.cs
Chocolatey.Cake.Recipe generates a SolutionVersion.cs file rather than
a SolutionInfo.cs file.

While we could have passed into a custom parameter to
Chocolatey.Cake.Recipe to control the generated file name, it was
decided to switch to using the default name, so that there was
consistency across the other projects that are using
Chocolatey.Cake.Recipe.