(#928) Update Chocolatey CLI to latest pre-release
This commit updates the the dependency for Chocolatey CLI
to the latest pre-release we have available internally.
This will prevent temporarily users from building the project
until a new release is made.

Additionally a pre-release version range is added to the nuspec
file to allow us to test upcoming functionality in the way we
want.