(maint) Add sources explicitly during cake build
This commit updates the cake build script to add
the sources we need to use explicitly, with the
ability to add a nuget source from an environment
variable when it is present.

This is needed due to a temporary change to use
versions of Chocolatey CLI that are only available
on our internal feed currently.