(#494) Refresh the sources from file each time
Sometimes (it looks like maybe when the Chocolatey Licensed Extension is
in play) we don't get the sources currently in the configuration file.
If instead of reusing the configuration settings service we instantiate
a new one each time, then it will ensure the configuration file is read
again. This resolves the issue of settings not applying in GUI if
they're made while GUI is running.