(#978) Always delete outdated cache file
Due to the way that the XmlService class works, if the information
about the outdated packages is the same as what is currently present in
the existing file, the file isn't updated. This is because the
XmlService compares the hash, and chooses to not update the file,
leaving it in tact.  Depending on the use case, this is exactly the
right thing to do, but for Chocolatey GUI, we use the date time of when
the file was created as a trigger as to whether another outdated check
is required.  As a result, multiple repeated outdated checks can happen
when not required.

This commit first deletes the cached file if it exists, and a new
outdated check was performed, writing a new file each time. This way,
the outdated check will only be performed once, when required.