(#977) Ensure that the correct public key is used
When attempting to load the Chocolatey GUI Licensed Extension assembly,
we need to ensure that the correct public key token is used.  In
previous versions of Chocolatey.Lib, the resolve_or_load_assembly
method made assumptions that meant that this wasn't necessary, but a
recent change over there has meant that it is now required.

Without passing in this information, the Chocolatey GUI Licensed
Extension assembly can't be located, and as a result, none of the
licensed functionality currently works within Chocolatey GUI.