(#974) Modify references to Chocolatey assembly
In order for things to work correctly, both during debugging, and
release builds, it is necessary to remove the explicit targetting of
the referenced Chocolatey assembly.  This is similar to what is done
elsewhere in other Chocolatey codebases, and allows for both officially
signed, and non-officially signed, refernces to be used.