(build) Add script to fetch latest Chocolatey bits
When debugging Chocolatey GUI with the latest Chocolatey library, it is
necessary to first build the Chocolatey CLI project, and copy the
generated assemblies into the correct location.  To save repeated work,
this script (similar ones are used in other repositories) does that
work for us, in a repeatable and configurable way.