(#980) Use assembly location when debugging
This sets the Chocolatey install directory to the current assembly
directory when built as a debug build. This allows correctly getting the
location of remembered arguments from the debug installed package(s)
instead of from the systemwide install of Chocolatey.

If a debug build is in use, then Chocolatey GUI is using the chocolatey
lib assembly, so true can be returned in all cases.