(#974) Add required properties to csproj files
In a previous commit:

https://github.com/chocolatey/ChocolateyGUI/commit/86f974c9e8c636d96a5330ea8c357f6e463425ad

The SpecificVersion and Private properties had been removed, which
meants that the chocolatey.dll was being copied, incorrectly, to the
output folder, as well as requiring a specificversion of a referenced
assembly to be provided.

The former has been "fixed" by ensuring that the MSI project doesn't
include the chocolatey.dll into the MSI, however, this commit addresses
the issue at the source as well.  This means that while in Debug mode,
the chocolatey.dll will be copied to the output folder to help with
debugging of Chocolatey GUI, however, in other build configurations, it
will not be copied.