(#989) Remove usage of side by side
With the upstream change in Chocolatey CLI to remove the usage of side
by side installations, we need to do the same thing within Chocolatey
GUI to prevent errors from happening.

This commit removes the usage of the properties that have been removed
from Chocolatey CLI, as well as changes the GUI to not include the
setting of these properties.