(build) Fix GitVersion support branch config
When the support branch configuration was created, it was copied from
the default configuration for develop/master branch.  In those
branches, we want to track release branches to know when there should
be a bump on version number.  However, a support branch is it's own
unique island of releases, and as such, we don't want/need to track
release branches.  The need for this change was found when attempting
to release 5.0.2, which was asserting the wrong version number due to
the presence of the release/6.0.0 branch.