(maint) Switch from int to long
When testing usage of nuget.org, there is a single package that has
over 3 billion downloads.  When attempting to view this package in
Chocolatey GUI, there is a convert error, when it tries to convert this
nubmer into an int.  This is a similar change to what was done in CCR,
when we crossed over the 1B mark for downloads.