(#999) Switch to using inclusive maximum version
We found that there were issues when using the exclusive maximum
version when it comes to having stable and pre-release package
versions, so the decision was taken to remove this variable by using a
very high package version number as the upper bound.  That way, when
the cross over between the major version happens, there will be no
conflict.

This commit simply changes the upper bound on the version range to use
the inclusive maximum version, ], rather than the exclusive one, ).