(build) Re-enable release notes export
Now that we are getting ready to ship the stable release of Chocolatey
GUI, we need to re-enable the release notes export.  There are still
two underlying issues with Chocolatey.Cake.Recipe where the export
doesn't work when it is an alpha/beta tag, and also when running on a
support branch, but neither of these should prevent the export when we
ship 2.0.0.