(#1000) Ensure sensitive arguments are detected
When the ability to show remembered arguments in Chocolatey GUI was
first introduced, code from the set_package_config_for_upgrade (which
was renamed to SetConfigFromRememberedArguments) method was borrowed,
as it was doing very similar work.  However, the code that was brought
over to Chocolatey GUI failed to realise that after splitting the
arguments on " --", when passed into the
arguments_contain_sensitive_information (which was renamed to
SensitiveArgumentsProvided) that a sensitive variable would no longer
be detected as this method expects that all arguments under test start
with a "-".

Without a major refactoring of how this works, the simplest way is to
put "--" back on the start of each argument, before passing into the
arguments_contain_sensitive_information method, that way, it will
correctly return whether it is a sensitive argument or not.

There is a larger discussion around whether a sensitive argument should
be persisted or not, but that is otwith the scope of this change.