(#1003) Handle null keys during source translation
The helper class that is used whe looking up any translations that has
been made did not handle the translation source strings when they
are null or empty. This causes problems when especially looking up
description of Chocolatey settings items that by some reason did
not have a description associated with themself in the configuration
file.