(build) Make build files consistent with others
The build entry scripts in this repository are slightly different to
the other repositories that we have, and there is no good reason for
them to be different.

This commits adds the .bat files from chocolatey/choco repository so
that they are consistent, and there aren't any surprises between how
builds are executed.