(#1014) Add advanced option to ignore cached items
This adds a new option to the advanced installation dialog
to be able to ignore any cached http resources that Chocolatey
CLI has cached. This can be useful for sources that allows an existing
version on the source can be replace, and then allow Chocolatey GUI
to download the absolute latest version of the package.