(maint) Remove version bump in nuspec file
The version bump in the nuspec file is being reverted, since this
shouldn't be required.  While there has been a bump in the csproj file
and the package.config files, the code related changes need to be done
in a backwards compatible way, so that newer versions of Chocolatey GUI
continue to work with older versions of Chocolatey CLI, etc.