(build) Modify how chained builds are executed
Now that Chocolatey CLI build is using Chocolatey.Cake.Recipe 0.24.0,
we have more control over which tasks as executed, and which aren't.
As a result, there is no need to invoke Cake twice, with different
entry points, but rather, simply prevent the execution of the tasks
that we don't want, using the command line options.