(maint) Allow better debugging of Chocolatey GUI
This commit introduces a change similar to what exists in Chocolatey
CLI.  While in debug mode, we source the installation location of both
Chocolatey and the GUI Licensed assembly from a different location,
rather than the official location.  This means that we can run
the debug version of Chocolatey GUI using a locally sourced debug
version of Chocolatey, which means that debugging some scenarios is
much easier.