(build) Alter signing steps
In the previous round of changes, we were a little too quick to remove
the signing step from the main build configuration.  This is needed, as
well as calling build.official.bat, in order to get an official
release.  The other build configurations, except the signing scripts,
one, don't need the signing keys in place, so this has also been done
in this commit.