(maint) Add CODEOWNERS files
This is an attempt to ensure that any changes to our dependencies,
whether that is:

* changes to NuGet packages through packages.config files

Are tracked and updated properly.  By making these files/folder owned
by the new @chocolatey-credits-reviewers group, it means that a member
of that group will be flagged on any PR's that we raised that contain
these files.